Не успел разобратьс со сборщиком, классы сделал но отваливается корзина
