+++ 3-STUDENTS/Demin-Konstantine/package.json
@@ -4,7 +4,7 @@
     "description": "",
     "main": "index.js",
     "scripts": {
-        "build": "webpack"
+        "build": "webpack --mode production"
     },
     "keywords": [],
     "author": "",
