+++ 3-STUDENTS/Demin-Konstantine/src/components/catalog.js
@@ -1,4 +1,3 @@
-
 function createItemTemplate(item) {
     return `<div class="catalog__item" id="${item.productId}">
     <button 
@@ -24,24 +23,29 @@ function createItemTemplate(item) {
 }
 
 
-let catalog = {
-    container: null,
-    url: 'https://raw.githubusercontent.com/kellolo/static/master/JSON/catalog.json',
-    items: [],
-    basket: null,
-    init () {
+export default class Catalog {
+    constructor() {
+        this.container = null
+        this.url = 'https://raw.githubusercontent.com/kellolo/static/master/JSON/catalog.json'
+        this.items = []
+        this.basket = null
+        this._init()
+    }
+
+    _init() {
         this.container = document.querySelector('#catalog');
         this.getData(this.url)
-            .then(items => {this.items = items})
+            .then(items => { this.items = items })
             .finally(() => {
                 this._render();
-                this.basket = basket; //ссылка на объект basket из файла cart.js
                 this.handleActions();
             })
-    },
+    }
+
     getData(url) {
         return fetch(url).then(data => data.json())
-    },
+    }
+
     handleActions() {
         this.container.addEventListener('click', evt => {
             if (evt.target.name == 'add') {
@@ -57,7 +61,7 @@ let catalog = {
                 this.basket.add(newProd);
             }
         })
-    },
+    }
     _render() {
         let htmlStr = '';
         this.items.forEach(item => {
@@ -66,5 +70,4 @@ let catalog = {
         this.container.innerHTML = htmlStr;
     }
 }
-
-
+const catalog = new Catalog();
