+++ 3-STUDENTS/Demin-Konstantine/src/components/index.js
@@ -1,0 +1,7 @@
+import basket from './basket.js'
+import catalog from './catalog.js'
+
+export default () => {
+    let basket = new Basket();
+    let catalog = new Catalog(basket);
+}
