+++ 3-STUDENTS/Demin-Konstantine/src/index.js
@@ -1,2 +1,6 @@
 import app from './components'
+// import basket from './basket'
+// import catalog from './catalog'
+// basket();
+// catalog();
 app();
